Fixing a type issue

+++ device_libraries/firehose_ipu.cpp
@@ -200,7 +200,7 @@ void tensorDecomp(long unsigned int row, long unsigned int col, long unsigned in
     std::vector<std::vector<float>> cpu_out1(num_streams);
 
     std::vector<float> temp_vec(row*col);
-    int value = 5; 
+    float value = 5.0; 
     fill(temp_vec.begin(), temp_vec.end(), value);
 
     for (int i = 0; i < num_streams; i++) {
